27.07.2012 - code style refactoring. javascript rhino testing

+++ CorePlugin/src/org/chaosfisch/youtubeuploader/plugins/coreplugin/modules/BindingsModule.java
@@ -30,15 +30,11 @@ import org.apache.ibatis.plugin.Interceptor;
 import org.apache.ibatis.session.*;
 import org.apache.ibatis.type.JdbcType;
 import org.apache.ibatis.type.TypeHandler;
+import org.apache.log4j.Logger;
 import org.apache.xbean.finder.ResourceFinder;
-import org.chaosfisch.youtubeuploader.plugins.coreplugin.services.impl.AccountServiceImpl;
-import org.chaosfisch.youtubeuploader.plugins.coreplugin.services.impl.PlaylistServiceImpl;
-import org.chaosfisch.youtubeuploader.plugins.coreplugin.services.impl.PresetServiceImpl;
-import org.chaosfisch.youtubeuploader.plugins.coreplugin.services.impl.QueueServiceImpl;
-import org.chaosfisch.youtubeuploader.plugins.coreplugin.services.spi.AccountService;
-import org.chaosfisch.youtubeuploader.plugins.coreplugin.services.spi.PlaylistService;
-import org.chaosfisch.youtubeuploader.plugins.coreplugin.services.spi.PresetService;
-import org.chaosfisch.youtubeuploader.plugins.coreplugin.services.spi.QueueService;
+import org.chaosfisch.youtubeuploader.plugins.coreplugin.CorePlugin;
+import org.chaosfisch.youtubeuploader.plugins.coreplugin.services.impl.*;
+import org.chaosfisch.youtubeuploader.plugins.coreplugin.services.spi.*;
 import org.chaosfisch.youtubeuploader.plugins.coreplugin.util.AutoTitleGeneratorImpl;
 import org.chaosfisch.youtubeuploader.plugins.coreplugin.util.spi.AutoTitleGenerator;
 import org.mybatis.guice.mappers.MapperProvider;
@@ -69,13 +65,17 @@ public class BindingsModule extends AbstractModule
 	protected void configure()
 	{
 
-		this.install(new CustomMyBatisModule()
+		install(new CustomMyBatisModule()
 		{
 			@Override protected void initialize()
 			{
 				final File outputFile = new File(String.format("%s/SimpleJavaYoutubeUploader/mybatis.xml", System.getProperty("user.home"))); //NON-NLS
 				try {
 					final File templateFile = new File("mybatis_config_template.xml");
+					Logger.getLogger(CorePlugin.class).info(String.format("Searching file mybatis template in %s", templateFile.getAbsolutePath()));
+					if (!templateFile.exists()) {
+						throw new RuntimeException("Required file \"mybatis_config_template.xml\" not found.");
+					}
 					final String fileContent = BindingsModule.readFileAsString(templateFile);
 
 					final FileWriter fstream = new FileWriter(outputFile);
@@ -102,28 +102,34 @@ public class BindingsModule extends AbstractModule
 						out.write(String.format(fileContent, typeAliases, mappers));
 					} catch (IOException ignored) {
 					} finally {
-						out.close();
-						fstream.close();
+						try {
+							out.close();
+							fstream.close();
+						} catch (IOException ignored) {
+						}
 					}
-					this.setClassPathResource(outputFile.getAbsolutePath());
-				} catch (FileNotFoundException ignored) {
-				} catch (IOException ignored) {
+					setClassPathResource(outputFile.getAbsolutePath());
+				} catch (FileNotFoundException e) {
+					e.printStackTrace();  //To change body of catch statement use File | Settings | File Templates.
+				} catch (IOException e) {
+					e.printStackTrace();  //To change body of catch statement use File | Settings | File Templates.
 				}
 			}
 		});
-		this.bind(AccountService.class).to(AccountServiceImpl.class).in(Singleton.class);
-		this.bind(QueueService.class).to(QueueServiceImpl.class).in(Singleton.class);
-		this.bind(PresetService.class).to(PresetServiceImpl.class).in(Singleton.class);
-		this.bind(PlaylistService.class).to(PlaylistServiceImpl.class).in(Singleton.class);
-		this.bind(AutoTitleGenerator.class).to(AutoTitleGeneratorImpl.class);
+		bind(AccountService.class).to(AccountServiceImpl.class).in(Singleton.class);
+		bind(QueueService.class).to(QueueServiceImpl.class).in(Singleton.class);
+		bind(PresetService.class).to(PresetServiceImpl.class).in(Singleton.class);
+		bind(PlaylistService.class).to(PlaylistServiceImpl.class).in(Singleton.class);
+		bind(PlaceholderService.class).to(PlaceholderServiceImpl.class).in(Singleton.class);
+		bind(AutoTitleGenerator.class).to(AutoTitleGeneratorImpl.class);
 	}
 
-	private static String readFileAsString(final File file) throws IOException, FileNotFoundException
+	private static String readFileAsString(final File file) throws IOException
 	{
 		final StringBuilder fileData = new StringBuilder(1000);
 		final BufferedReader reader = new BufferedReader(new FileReader(file));
 		char[] buf = new char[1024];
-		int numRead = 0;
+		int numRead;
 		while ((numRead = reader.read(buf)) != -1) {
 			final String readData = String.valueOf(buf, 0, numRead);
 			fileData.append(readData);
@@ -136,29 +142,29 @@ public class BindingsModule extends AbstractModule
 	private abstract static class AbstractMyBatisModule extends AbstractModule
 	{
 
-		private ClassLoader resourcesClassLoader = this.getDefaultClassLoader();
+		private ClassLoader resourcesClassLoader = getDefaultClassLoader();
 
-		private ClassLoader driverClassLoader = this.getDefaultClassLoader();
+		private ClassLoader driverClassLoader = getDefaultClassLoader();
 
 		@Override
 		protected final void configure()
 		{
 			try {
 				// sql session manager
-				this.bind(SqlSessionManager.class).toProvider(SqlSessionManagerProvider.class).in(Scopes.SINGLETON);
-				this.bind(SqlSession.class).to(SqlSessionManager.class).in(Scopes.SINGLETON);
+				bind(SqlSessionManager.class).toProvider(SqlSessionManagerProvider.class).in(Scopes.SINGLETON);
+				bind(SqlSession.class).to(SqlSessionManager.class).in(Scopes.SINGLETON);
 
 				// transactional interceptor
 				final TransactionalMethodInterceptor interceptor = new TransactionalMethodInterceptor();
-				this.requestInjection(interceptor);
-				this.bindInterceptor(any(), annotatedWith(Transactional.class), interceptor);
+				requestInjection(interceptor);
+				bindInterceptor(any(), annotatedWith(Transactional.class), interceptor);
 
-				this.internalConfigure();
+				internalConfigure();
 
-				this.bind(ClassLoader.class).annotatedWith(named("JDBC.driverClassLoader")).toInstance(this.driverClassLoader); //NON-NLS
+				bind(ClassLoader.class).annotatedWith(named("JDBC.driverClassLoader")).toInstance(driverClassLoader); //NON-NLS
 			} finally {
-				this.resourcesClassLoader = this.getDefaultClassLoader();
-				this.driverClassLoader = this.getDefaultClassLoader();
+				resourcesClassLoader = getDefaultClassLoader();
+				driverClassLoader = getDefaultClassLoader();
 			}
 		}
 
@@ -168,7 +174,7 @@ public class BindingsModule extends AbstractModule
 		 */
 		final <T> void bindMapper(final Class<T> mapperType)
 		{
-			this.bind(mapperType).toProvider(guicify(new MapperProvider<T>(mapperType))).in(Scopes.SINGLETON);
+			bind(mapperType).toProvider(guicify(new MapperProvider<T>(mapperType))).in(Scopes.SINGLETON);
 		}
 
 		/**
@@ -176,7 +182,7 @@ public class BindingsModule extends AbstractModule
 		 */
 		public void useResourceClassLoader(final ClassLoader resourceClassLoader)
 		{
-			this.resourcesClassLoader = resourceClassLoader;
+			resourcesClassLoader = resourceClassLoader;
 		}
 
 		/**
@@ -184,7 +190,7 @@ public class BindingsModule extends AbstractModule
 		 */
 		protected final ClassLoader getResourceClassLoader()
 		{
-			return this.resourcesClassLoader;
+			return resourcesClassLoader;
 		}
 
 		/**
@@ -200,7 +206,7 @@ public class BindingsModule extends AbstractModule
 		 */
 		private ClassLoader getDefaultClassLoader()
 		{
-			return this.getClass().getClassLoader();
+			return getClass().getClassLoader();
 		}
 
 		/**
@@ -270,13 +276,13 @@ public class BindingsModule extends AbstractModule
 		@Override
 		final void internalConfigure()
 		{
-			this.initialize();
+			initialize();
 
 			Reader reader = null;
 			try {
-				reader = new FileReader(this.classPathResource);
-				final SqlSessionFactory sessionFactory = new SqlSessionFactoryBuilder().build(reader, this.environmentId, this.properties);
-				this.bind(SqlSessionFactory.class).toInstance(sessionFactory);
+				reader = new FileReader(classPathResource);
+				final SqlSessionFactory sessionFactory = new SqlSessionFactoryBuilder().build(reader, environmentId, properties);
+				bind(SqlSessionFactory.class).toInstance(sessionFactory);
 
 				final Configuration configuration = sessionFactory.getConfiguration();
 
@@ -287,32 +293,32 @@ public class BindingsModule extends AbstractModule
 				// bind mappers
 				@SuppressWarnings("unchecked") final Iterable<Class<?>> mapperClasses = (Iterable<Class<?>>) getValue(CustomMyBatisModule.KNOWN_MAPPERS, context, configuration);
 				for (final Class<?> mapperType : mapperClasses) {
-					this.bindMapper(mapperType);
+					bindMapper(mapperType);
 				}
 
 				// request injection for type handlers
 				@SuppressWarnings("unchecked") final Iterable<Map<JdbcType, TypeHandler<?>>> mappedTypeHandlers = (Iterable<Map<JdbcType, TypeHandler<?>>>) getValue(CustomMyBatisModule.TYPE_HANDLERS, context, configuration);
 				for (final Map<JdbcType, TypeHandler<?>> mappedTypeHandler : mappedTypeHandlers) {
 					for (final TypeHandler<?> handler : mappedTypeHandler.values()) {
-						this.requestInjection(handler);
+						requestInjection(handler);
 					}
 				}
 				@SuppressWarnings("unchecked") final Iterable<TypeHandler<?>> allTypeHandlers = (Iterable<TypeHandler<?>>) getValue(CustomMyBatisModule.ALL_TYPE_HANDLERS, context, configuration);
 				for (final TypeHandler<?> handler : allTypeHandlers) {
-					this.requestInjection(handler);
+					requestInjection(handler);
 				}
 
 				// request injection for interceptors
 				@SuppressWarnings("unchecked") final Iterable<Interceptor> interceptors = (Iterable<Interceptor>) getValue(CustomMyBatisModule.INTERCEPTORS, context, configuration);
 				for (final Interceptor interceptor : interceptors) {
-					this.requestInjection(interceptor);
+					requestInjection(interceptor);
 				}
 			} catch (FileNotFoundException e) {
 				//noinspection DuplicateStringLiteralInspection
-				this.addError("Impossible to read classpath resource '%s', see nested exceptions: %s", this.classPathResource, e.getMessage()); //NON-NLS
+				addError("Impossible to read classpath resource '%s', see nested exceptions: %s", classPathResource, e.getMessage()); //NON-NLS
 			} catch (OgnlException e) {
 				//noinspection DuplicateStringLiteralInspection
-				this.addError("Impossible to read classpath resource '%s', see nested exceptions: %s", this.classPathResource, e.getMessage()); //NON-NLS
+				addError("Impossible to read classpath resource '%s', see nested exceptions: %s", classPathResource, e.getMessage()); //NON-NLS
 			} finally {
 				if (reader != null) {
 					try {
